Squash: SystemUI: Tile: Lockscreen fixed for 7.0 by @Millosr
SystemUI: Tile: Lockscreen: Fix lockscreen bypass
Disable the Lockscreen tile when we are in the lockscreen and
when this one is securized.

That avoid to bypass the lockscreen (security issue)

This replace state.visible unsupported since 7.0 and
avoid any confusion for the end user because the tile is
still visible but disabled.

SystemUI: improve lockscreen tile behavior by romanbb
* Stay disabled when disabled
* No need to use Keyguard locks, use internal update monitor