Squash of v20 second screen offset commits display: Allow devices to manually offset window center
Some device have irrugularly shaped screens or screens apended to them
that we do not use, however they are treated as part of the main screen.
Allow devices to specify, via overlay, offset values to pad the center
of the screen.

display: Honor display offset when rotating screen

Currently if an offset is set and the screen is rotated, windowmanager
continues to offset in the direction (up/down, left/right) that the offset
was originally set to. Make it honor rotation and change the offset
direction on rotation.

Change-Id: I0e3f830d468b6c75c238c92e73366e0fa2def11e